cmd: Add trace subcommand.
This subcommand groups all *snoop gadgets like execsnoop or mountsnoop.
