+++ cmd/kubectl-gadget/trace/bind.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package trace
 
 import (
 	"encoding/json"
@@ -22,45 +22,50 @@ import (
 	"strings"
 
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
-	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/sigsnoop/types"
+	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/bindsnoop/types"
 	eventtypes "github.com/kinvolk/inspektor-gadget/pkg/types"
 	"github.com/spf13/cobra"
 )
 
 var (
-	pid    uint
-	sig    string
-	failed bool
+	targetPid    uint
+	targetPorts  []uint
+	ignoreErrors bool
 )
 
-var sigsnoopCmd = &cobra.Command{
-	Use:   "sigsnoop",
-	Short: "Trace signals received by processes",
+var bindsnoopCmd = &cobra.Command{
+	Use:   "bind",
+	Short: "Trace the kernel functions performing socket binding",
 	RunE: func(cmd *cobra.Command, args []string) error {
+		// print header
 		switch params.OutputMode {
-		case utils.OutputModeJSON: // don't print any header
 		case utils.OutputModeCustomColumns:
-			fmt.Println(getCustomSigsnoopColsHeader(params.CustomColumns))
+			fmt.Println(getCustomBindsnoopColsHeader(params.CustomColumns))
 		case utils.OutputModeColumns:
-			fmt.Printf("%-16s %-16s %-16s %-16s %-6s %-16s %-9s %-6s %-6s\n",
+			fmt.Printf("%-16s %-16s %-16s %-16s %-6s %-16s %-6s %-16s %-6s %-6s %s\n",
 				"NODE", "NAMESPACE", "POD", "CONTAINER",
-				"PID", "COMM", "SIGNAL", "TPID", "RET")
+				"PID", "COMM", "PROTO", "ADDR", "PORT", "OPTS", "IF")
+		}
+
+		portsStringSlice := []string{}
+		for _, port := range targetPorts {
+			portsStringSlice = append(portsStringSlice, strconv.FormatUint(uint64(port), 10))
 		}
 
 		config := &utils.TraceConfig{
-			GadgetName:       "sigsnoop",
+			GadgetName:       "bindsnoop",
 			Operation:        "start",
 			TraceOutputMode:  "Stream",
 			TraceOutputState: "Started",
 			CommonFlags:      &params,
 			Parameters: map[string]string{
-				"signal": sig,
-				"pid":    strconv.FormatUint(uint64(pid), 10),
-				"failed": strconv.FormatBool(failed),
+				"pid":           strconv.FormatUint(uint64(targetPid), 10),
+				"ports":         strings.Join(portsStringSlice, ","),
+				"ignore_errors": strconv.FormatBool(ignoreErrors),
 			},
 		}
 
-		err := utils.RunTraceAndPrintStream(config, sigsnoopTransformLine)
+		err := utils.RunTraceAndPrintStream(config, bindsnoopTransformLine)
 		if err != nil {
 			return utils.WrapInErrRunGadget(err)
 		}
@@ -70,33 +75,35 @@ var sigsnoopCmd = &cobra.Command{
 }
 
 func init() {
-	rootCmd.AddCommand(sigsnoopCmd)
-	utils.AddCommonFlags(sigsnoopCmd, &params)
+	TraceCmd.AddCommand(bindsnoopCmd)
+	utils.AddCommonFlags(bindsnoopCmd, &params)
 
-	sigsnoopCmd.PersistentFlags().UintVarP(
-		&pid,
+	bindsnoopCmd.PersistentFlags().UintVarP(
+		&targetPid,
 		"pid",
 		"",
 		0,
-		"Show only signal sent by this particular PID",
+		"Show only bind events generated by this particular PID",
 	)
-	sigsnoopCmd.PersistentFlags().StringVarP(
-		&sig,
-		"signal",
-		"",
-		"",
-		`Trace only this signal (it can be an int like 9 or string beginning with "SIG" like "SIGKILL")`,
+	bindsnoopCmd.PersistentFlags().UintSliceVarP(
+		&targetPorts,
+		"ports",
+		"P",
+		[]uint{},
+		"Trace only bind events involving these ports",
 	)
-	sigsnoopCmd.PersistentFlags().BoolVarP(
-		&failed,
-		"failed-only",
-		"f",
-		false,
-		`Show only events where the syscall sending a signal failed`,
+	bindsnoopCmd.PersistentFlags().BoolVarP(
+		&ignoreErrors,
+		"ignore-errors",
+		"i",
+		true,
+		"Show only events where the bind succeeded",
 	)
 }
 
-func sigsnoopTransformLine(line string) string {
+// bindsnoopTransformLine is called to transform an event to columns
+// format according to the parameters
+func bindsnoopTransformLine(line string) string {
 	var sb strings.Builder
 	var e types.Event
 
@@ -107,19 +114,18 @@ func sigsnoopTransformLine(line string) string {
 
 	if e.Type == eventtypes.ERR || e.Type == eventtypes.WARN ||
 		e.Type == eventtypes.DEBUG || e.Type == eventtypes.INFO {
-		fmt.Fprintf(os.Stderr, "%s: node %q: %s", e.Type, e.Node, e.Message)
+		fmt.Fprintf(os.Stderr, "%s: node %s: %s", e.Type, e.Node, e.Message)
 		return ""
 	}
 
 	if e.Type != eventtypes.NORMAL {
 		return ""
 	}
-
 	switch params.OutputMode {
 	case utils.OutputModeColumns:
-		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-6d %-16s %-9s %-6d %-6d",
-			e.Node, e.Namespace, e.Pod, e.Container, e.Pid, e.Comm,
-			e.Signal, e.TargetPid, e.Retval))
+		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-6d %-16s %-6s %-16s %-6d %-6s %s",
+			e.Node, e.Namespace, e.Pod, e.Container,
+			e.Pid, e.Comm, e.Protocol, e.Addr, e.Port, e.Options, e.Interface))
 	case utils.OutputModeCustomColumns:
 		for _, col := range params.CustomColumns {
 			switch col {
@@ -135,12 +141,16 @@ func sigsnoopTransformLine(line string) string {
 				sb.WriteString(fmt.Sprintf("%-6d", e.Pid))
 			case "comm":
 				sb.WriteString(fmt.Sprintf("%-16s", e.Comm))
-			case "signal":
-				sb.WriteString(fmt.Sprintf("%-9s", e.Signal))
-			case "tpid":
-				sb.WriteString(fmt.Sprintf("%-6d", e.TargetPid))
-			case "ret":
-				sb.WriteString(fmt.Sprintf("%-6d", e.Retval))
+			case "proto":
+				sb.WriteString(fmt.Sprintf("%-6s", e.Protocol))
+			case "addr":
+				sb.WriteString(fmt.Sprintf("%-16s", e.Addr))
+			case "port":
+				sb.WriteString(fmt.Sprintf("%-6d", e.Port))
+			case "opts":
+				sb.WriteString(fmt.Sprintf("%-6s", e.Options))
+			case "if":
+				sb.WriteString(fmt.Sprintf("%-6s", e.Interface))
 			}
 			sb.WriteRune(' ')
 		}
@@ -149,7 +159,7 @@ func sigsnoopTransformLine(line string) string {
 	return sb.String()
 }
 
-func getCustomSigsnoopColsHeader(cols []string) string {
+func getCustomBindsnoopColsHeader(cols []string) string {
 	var sb strings.Builder
 
 	for _, col := range cols {
@@ -166,12 +176,16 @@ func getCustomSigsnoopColsHeader(cols []string) string {
 			sb.WriteString(fmt.Sprintf("%-6s", "PID"))
 		case "comm":
 			sb.WriteString(fmt.Sprintf("%-16s", "COMM"))
-		case "signal":
-			sb.WriteString(fmt.Sprintf("%-9s", "SIGNAL"))
-		case "tpid":
-			sb.WriteString(fmt.Sprintf("%-6s", "TPID"))
-		case "ret":
-			sb.WriteString(fmt.Sprintf("%-6s", "RET"))
+		case "proto":
+			sb.WriteString(fmt.Sprintf("%-6s", "PROTO"))
+		case "addr":
+			sb.WriteString(fmt.Sprintf("%-16s", "ADDR"))
+		case "port":
+			sb.WriteString(fmt.Sprintf("%-6s", "PORT"))
+		case "opts":
+			sb.WriteString(fmt.Sprintf("%-6s", "OPTS"))
+		case "if":
+			sb.WriteString(fmt.Sprintf("%-6s", "IF"))
 		}
 		sb.WriteRune(' ')
 	}
