+++ cmd/kubectl-gadget/trace/capabilities.go
@@ -12,78 +12,43 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package trace
 
 import (
 	"encoding/json"
 	"fmt"
-	"math"
 	"os"
-	"strconv"
 	"strings"
 
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
-	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/fsslower/types"
+	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/capabilities/types"
 	eventtypes "github.com/kinvolk/inspektor-gadget/pkg/types"
 	"github.com/spf13/cobra"
 )
 
-var (
-	// flags
-	fsslowerMinLatency uint
-	fsslowerFilesystem string
-)
-
-var validFsSlowerFilesystems = []string{"btrfs", "ext4", "nfs", "xfs"}
-
-var fsslowerCmd = &cobra.Command{
-	Use:   "fsslower",
-	Short: "Trace open, read, write and fsync operations slower than a threshold",
-	PreRunE: func(cmd *cobra.Command, args []string) error {
-		if fsslowerFilesystem == "" {
-			return utils.WrapInErrMissingArgs("--filesystem")
-		}
-
-		found := false
-
-		for _, val := range validFsSlowerFilesystems {
-			if fsslowerFilesystem == val {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			return utils.WrapInErrInvalidArg("--type / -t",
-				fmt.Errorf("%q is not a valid filesystem", fsslowerFilesystem))
-		}
-
-		return nil
-	},
+var capabilitiesCmd = &cobra.Command{
+	Use:   "capabilities",
+	Short: "Trace security capability checks",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// print header
 		switch params.OutputMode {
 		case utils.OutputModeCustomColumns:
-			fmt.Println(getCustomFsslowerColsHeader(params.CustomColumns))
+			fmt.Println(getCustomOpensnoopColsHeader(params.CustomColumns))
 		case utils.OutputModeColumns:
-			fmt.Printf("%-16s %-16s %-16s %-16s %-16s %-6s %1s %-6s %-7s %-8s %s\n",
+			fmt.Printf("%-16s %-16s %-16s %-16s %-6s %-6s %-16s %-4s %-16s %-6s\n",
 				"NODE", "NAMESPACE", "POD", "CONTAINER",
-				"COMM", "PID", "T", "BYTES", "OFFSET", "LAT(ms)", "FILE")
+				"UID", "PID", "COMM", "CAP", "NAME", "AUDIT")
 		}
 
 		config := &utils.TraceConfig{
-			GadgetName:       "fsslower",
+			GadgetName:       "capabilities",
 			Operation:        "start",
 			TraceOutputMode:  "Stream",
 			TraceOutputState: "Started",
 			CommonFlags:      &params,
-			Parameters: map[string]string{
-				"filesystem": fsslowerFilesystem,
-				"minlatency": strconv.FormatUint(uint64(fsslowerMinLatency), 10),
-			},
 		}
 
-		err := utils.RunTraceAndPrintStream(config, fsslowerTransformLine)
+		err := utils.RunTraceAndPrintStream(config, capabilitiesTransformLine)
 		if err != nil {
 			return utils.WrapInErrRunGadget(err)
 		}
@@ -93,20 +58,13 @@ var fsslowerCmd = &cobra.Command{
 }
 
 func init() {
-	fsslowerCmd.Flags().UintVarP(
-		&fsslowerMinLatency, "min", "m", types.MinLatencyDefault,
-		"Min latency to trace, in ms",
-	)
-	fsslowerCmd.Flags().StringVarP(
-		&fsslowerFilesystem, "type", "t", "",
-		fmt.Sprintf("Which filesystem to trace: [%s]", strings.Join(validFsSlowerFilesystems, ", ")),
-	)
-
-	rootCmd.AddCommand(fsslowerCmd)
-	utils.AddCommonFlags(fsslowerCmd, &params)
+	TraceCmd.AddCommand(capabilitiesCmd)
+	utils.AddCommonFlags(capabilitiesCmd, &params)
 }
 
-func fsslowerTransformLine(line string) string {
+// capabilitiesTransformLine is called to transform an event to columns
+// format according to the parameters
+func capabilitiesTransformLine(line string) string {
 	var sb strings.Builder
 	var e types.Event
 
@@ -125,16 +83,11 @@ func fsslowerTransformLine(line string) string {
 		return ""
 	}
 
-	// TODO: what to print in this case?
-	if e.Bytes == math.MaxInt64 {
-		e.Bytes = 0
-	}
-
 	switch params.OutputMode {
 	case utils.OutputModeColumns:
-		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-16s %-6d %1s %-6d %-7d %-8.2f %s",
+		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-6d %-6d %-16s %-4d %-16s %-6d",
 			e.Node, e.Namespace, e.Pod, e.Container,
-			e.Comm, e.Pid, e.Op, e.Bytes, e.Offset, float64(e.Latency)/1000.0, e.File))
+			e.UID, e.Pid, e.Comm, e.Cap, e.CapName, e.Audit))
 	case utils.OutputModeCustomColumns:
 		for _, col := range params.CustomColumns {
 			switch col {
@@ -146,20 +99,18 @@ func fsslowerTransformLine(line string) string {
 				sb.WriteString(fmt.Sprintf("%-16s", e.Pod))
 			case "container":
 				sb.WriteString(fmt.Sprintf("%-16s", e.Container))
-			case "comm":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Comm))
 			case "pid":
 				sb.WriteString(fmt.Sprintf("%-6d", e.Pid))
-			case "t":
-				sb.WriteString(fmt.Sprintf("%1s", e.Op))
-			case "bytes":
-				sb.WriteString(fmt.Sprintf("%-6d", e.Bytes))
-			case "offset":
-				sb.WriteString(fmt.Sprintf("%-7d", e.Offset))
-			case "lat":
-				sb.WriteString(fmt.Sprintf("%-8.2f", float64(e.Latency)/1000.0))
-			case "file":
-				sb.WriteString(fmt.Sprintf("%24s", e.File))
+			case "uid":
+				sb.WriteString(fmt.Sprintf("%-6d", e.UID))
+			case "comm":
+				sb.WriteString(fmt.Sprintf("%-16s", e.Comm))
+			case "cap":
+				sb.WriteString(fmt.Sprintf("%-4d", e.Cap))
+			case "name":
+				sb.WriteString(fmt.Sprintf("%-16s", e.CapName))
+			case "audit":
+				sb.WriteString(fmt.Sprintf("%-6d", e.Audit))
 			}
 			sb.WriteRune(' ')
 		}
@@ -168,7 +119,7 @@ func fsslowerTransformLine(line string) string {
 	return sb.String()
 }
 
-func getCustomFsslowerColsHeader(cols []string) string {
+func getCustomCapabilitiesColsHeader(cols []string) string {
 	var sb strings.Builder
 
 	for _, col := range cols {
@@ -181,20 +132,18 @@ func getCustomFsslowerColsHeader(cols []string) string {
 			sb.WriteString(fmt.Sprintf("%-16s", "POD"))
 		case "container":
 			sb.WriteString(fmt.Sprintf("%-16s", "CONTAINER"))
-		case "comm":
-			sb.WriteString(fmt.Sprintf("%-16s", "COMM"))
 		case "pid":
 			sb.WriteString(fmt.Sprintf("%-6s", "PID"))
-		case "t":
-			sb.WriteString(fmt.Sprintf("%1s", "T"))
-		case "bytes":
-			sb.WriteString(fmt.Sprintf("%-6s", "BYTES"))
-		case "offset":
-			sb.WriteString(fmt.Sprintf("%-7s", "OFFSET"))
-		case "lat":
-			sb.WriteString(fmt.Sprintf("%-8s", "LAT(ms)"))
-		case "file":
-			sb.WriteString(fmt.Sprintf("%24s", "FILE"))
+		case "uid":
+			sb.WriteString(fmt.Sprintf("%-6s", "UID"))
+		case "comm":
+			sb.WriteString(fmt.Sprintf("%-16s", "COMM"))
+		case "cap":
+			sb.WriteString(fmt.Sprintf("%-4s", "CAP"))
+		case "name":
+			sb.WriteString(fmt.Sprintf("%-16s", "NAME"))
+		case "audit":
+			sb.WriteString(fmt.Sprintf("%-6s", "AUDIT"))
 		}
 		sb.WriteRune(' ')
 	}
