+++ cmd/kubectl-gadget/trace/fsslower.go
@@ -12,43 +12,78 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package trace
 
 import (
 	"encoding/json"
 	"fmt"
+	"math"
 	"os"
+	"strconv"
 	"strings"
 
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
-	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/tcpconnect/types"
+	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/fsslower/types"
 	eventtypes "github.com/kinvolk/inspektor-gadget/pkg/types"
 	"github.com/spf13/cobra"
 )
 
-var tcpconnectCmd = &cobra.Command{
-	Use:   "tcpconnect",
-	Short: "Trace connect() system calls",
+var (
+	// flags
+	fsslowerMinLatency uint
+	fsslowerFilesystem string
+)
+
+var validFsSlowerFilesystems = []string{"btrfs", "ext4", "nfs", "xfs"}
+
+var fsslowerCmd = &cobra.Command{
+	Use:   "fsslower",
+	Short: "Trace open, read, write and fsync operations slower than a threshold",
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		if fsslowerFilesystem == "" {
+			return utils.WrapInErrMissingArgs("--filesystem")
+		}
+
+		found := false
+
+		for _, val := range validFsSlowerFilesystems {
+			if fsslowerFilesystem == val {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			return utils.WrapInErrInvalidArg("--type / -t",
+				fmt.Errorf("%q is not a valid filesystem", fsslowerFilesystem))
+		}
+
+		return nil
+	},
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// print header
 		switch params.OutputMode {
 		case utils.OutputModeCustomColumns:
-			fmt.Println(getCustomTcpconnectColsHeader(params.CustomColumns))
+			fmt.Println(getCustomFsslowerColsHeader(params.CustomColumns))
 		case utils.OutputModeColumns:
-			fmt.Printf("%-16s %-16s %-16s %-16s %-6s %-16s %-3s %-16s %-16s %-7s\n",
+			fmt.Printf("%-16s %-16s %-16s %-16s %-16s %-6s %1s %-6s %-7s %-8s %s\n",
 				"NODE", "NAMESPACE", "POD", "CONTAINER",
-				"PID", "COMM", "IP", "SADDR", "DADDR", "DPORT")
+				"COMM", "PID", "T", "BYTES", "OFFSET", "LAT(ms)", "FILE")
 		}
 
 		config := &utils.TraceConfig{
-			GadgetName:       "tcpconnect",
+			GadgetName:       "fsslower",
 			Operation:        "start",
 			TraceOutputMode:  "Stream",
 			TraceOutputState: "Started",
 			CommonFlags:      &params,
+			Parameters: map[string]string{
+				"filesystem": fsslowerFilesystem,
+				"minlatency": strconv.FormatUint(uint64(fsslowerMinLatency), 10),
+			},
 		}
 
-		err := utils.RunTraceAndPrintStream(config, tcpconnectTransformLine)
+		err := utils.RunTraceAndPrintStream(config, fsslowerTransformLine)
 		if err != nil {
 			return utils.WrapInErrRunGadget(err)
 		}
@@ -58,13 +93,20 @@ var tcpconnectCmd = &cobra.Command{
 }
 
 func init() {
-	rootCmd.AddCommand(tcpconnectCmd)
-	utils.AddCommonFlags(tcpconnectCmd, &params)
+	fsslowerCmd.Flags().UintVarP(
+		&fsslowerMinLatency, "min", "m", types.MinLatencyDefault,
+		"Min latency to trace, in ms",
+	)
+	fsslowerCmd.Flags().StringVarP(
+		&fsslowerFilesystem, "type", "t", "",
+		fmt.Sprintf("Which filesystem to trace: [%s]", strings.Join(validFsSlowerFilesystems, ", ")),
+	)
+
+	TraceCmd.AddCommand(fsslowerCmd)
+	utils.AddCommonFlags(fsslowerCmd, &params)
 }
 
-// tcpconnectTransformLine is called to transform an event to columns
-// format according to the parameters
-func tcpconnectTransformLine(line string) string {
+func fsslowerTransformLine(line string) string {
 	var sb strings.Builder
 	var e types.Event
 
@@ -83,11 +125,16 @@ func tcpconnectTransformLine(line string) string {
 		return ""
 	}
 
+	// TODO: what to print in this case?
+	if e.Bytes == math.MaxInt64 {
+		e.Bytes = 0
+	}
+
 	switch params.OutputMode {
 	case utils.OutputModeColumns:
-		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-6d %-16s %-3d %-16s %-16s %-7d",
+		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-16s %-6d %1s %-6d %-7d %-8.2f %s",
 			e.Node, e.Namespace, e.Pod, e.Container,
-			e.Pid, e.Comm, e.IPVersion, e.Saddr, e.Daddr, e.Dport))
+			e.Comm, e.Pid, e.Op, e.Bytes, e.Offset, float64(e.Latency)/1000.0, e.File))
 	case utils.OutputModeCustomColumns:
 		for _, col := range params.CustomColumns {
 			switch col {
@@ -99,18 +146,20 @@ func tcpconnectTransformLine(line string) string {
 				sb.WriteString(fmt.Sprintf("%-16s", e.Pod))
 			case "container":
 				sb.WriteString(fmt.Sprintf("%-16s", e.Container))
-			case "pid":
-				sb.WriteString(fmt.Sprintf("%-6d", e.Pid))
 			case "comm":
 				sb.WriteString(fmt.Sprintf("%-16s", e.Comm))
-			case "ip":
-				sb.WriteString(fmt.Sprintf("%-3d", e.IPVersion))
-			case "saddr":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Saddr))
-			case "daddr":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Daddr))
-			case "dport":
-				sb.WriteString(fmt.Sprintf("%-7d", e.Dport))
+			case "pid":
+				sb.WriteString(fmt.Sprintf("%-6d", e.Pid))
+			case "t":
+				sb.WriteString(fmt.Sprintf("%1s", e.Op))
+			case "bytes":
+				sb.WriteString(fmt.Sprintf("%-6d", e.Bytes))
+			case "offset":
+				sb.WriteString(fmt.Sprintf("%-7d", e.Offset))
+			case "lat":
+				sb.WriteString(fmt.Sprintf("%-8.2f", float64(e.Latency)/1000.0))
+			case "file":
+				sb.WriteString(fmt.Sprintf("%24s", e.File))
 			}
 			sb.WriteRune(' ')
 		}
@@ -119,7 +168,7 @@ func tcpconnectTransformLine(line string) string {
 	return sb.String()
 }
 
-func getCustomTcpconnectColsHeader(cols []string) string {
+func getCustomFsslowerColsHeader(cols []string) string {
 	var sb strings.Builder
 
 	for _, col := range cols {
@@ -132,18 +181,20 @@ func getCustomTcpconnectColsHeader(cols []string) string {
 			sb.WriteString(fmt.Sprintf("%-16s", "POD"))
 		case "container":
 			sb.WriteString(fmt.Sprintf("%-16s", "CONTAINER"))
-		case "pid":
-			sb.WriteString(fmt.Sprintf("%-6s", "PID"))
 		case "comm":
 			sb.WriteString(fmt.Sprintf("%-16s", "COMM"))
-		case "ip":
-			sb.WriteString(fmt.Sprintf("%-3s", "IP"))
-		case "saddr":
-			sb.WriteString(fmt.Sprintf("%-16s", "SADDR"))
-		case "daddr":
-			sb.WriteString(fmt.Sprintf("%-16s", "DADDR"))
-		case "dport":
-			sb.WriteString(fmt.Sprintf("%-7s", "DPORT"))
+		case "pid":
+			sb.WriteString(fmt.Sprintf("%-6s", "PID"))
+		case "t":
+			sb.WriteString(fmt.Sprintf("%1s", "T"))
+		case "bytes":
+			sb.WriteString(fmt.Sprintf("%-6s", "BYTES"))
+		case "offset":
+			sb.WriteString(fmt.Sprintf("%-7s", "OFFSET"))
+		case "lat":
+			sb.WriteString(fmt.Sprintf("%-8s", "LAT(ms)"))
+		case "file":
+			sb.WriteString(fmt.Sprintf("%24s", "FILE"))
 		}
 		sb.WriteRune(' ')
 	}
