+++ cmd/kubectl-gadget/trace/open.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package trace
 
 import (
 	"encoding/json"
@@ -21,34 +21,34 @@ import (
 	"strings"
 
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
-	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/mountsnoop/types"
+	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/opensnoop/types"
 	eventtypes "github.com/kinvolk/inspektor-gadget/pkg/types"
 	"github.com/spf13/cobra"
 )
 
-var mountsnoopCmd = &cobra.Command{
-	Use:   "mountsnoop",
-	Short: "Trace open() system calls",
+var opensnoopCmd = &cobra.Command{
+	Use:   "open",
+	Short: "Trace open system calls",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// print header
 		switch params.OutputMode {
 		case utils.OutputModeCustomColumns:
-			fmt.Println(getCustomMountsnoopColsHeader(params.CustomColumns))
+			fmt.Println(getCustomOpensnoopColsHeader(params.CustomColumns))
 		case utils.OutputModeColumns:
-			fmt.Printf("%-16s %-16s %-16s %-16s %-16s %-6s %-6s %-10s %s\n",
+			fmt.Printf("%-16s %-16s %-16s %-16s %-6s %-16s %-3s %3s %s\n",
 				"NODE", "NAMESPACE", "POD", "CONTAINER",
-				"COMM", "PID", "TID", "MNT_NS", "CALL")
+				"PID", "COMM", "FD", "ERR", "PATH")
 		}
 
 		config := &utils.TraceConfig{
-			GadgetName:       "mountsnoop",
+			GadgetName:       "opensnoop",
 			Operation:        "start",
 			TraceOutputMode:  "Stream",
 			TraceOutputState: "Started",
 			CommonFlags:      &params,
 		}
 
-		err := utils.RunTraceAndPrintStream(config, mountsnoopTransformLine)
+		err := utils.RunTraceAndPrintStream(config, opensnoopTransformLine)
 		if err != nil {
 			return utils.WrapInErrRunGadget(err)
 		}
@@ -58,27 +58,13 @@ var mountsnoopCmd = &cobra.Command{
 }
 
 func init() {
-	rootCmd.AddCommand(mountsnoopCmd)
-	utils.AddCommonFlags(mountsnoopCmd, &params)
+	TraceCmd.AddCommand(opensnoopCmd)
+	utils.AddCommonFlags(opensnoopCmd, &params)
 }
 
-func getCall(e *types.Event) string {
-	switch e.Operation {
-	case "mount":
-		format := `mount("%s", "%s", "%s", %s, "%s") = %d`
-		return fmt.Sprintf(format, e.Source, e.Target, e.Fs, strings.Join(e.Flags, " | "),
-			e.Data, e.Retval)
-	case "umount":
-		format := `umount("%s", %s) = %d`
-		return fmt.Sprintf(format, e.Target, strings.Join(e.Flags, " | "), e.Retval)
-	}
-
-	return ""
-}
-
-// mountsnoopTransformLine is called to transform an event to columns
+// opensnoopTransformLine is called to transform an event to columns
 // format according to the parameters
-func mountsnoopTransformLine(line string) string {
+func opensnoopTransformLine(line string) string {
 	var sb strings.Builder
 	var e types.Event
 
@@ -99,9 +85,9 @@ func mountsnoopTransformLine(line string) string {
 
 	switch params.OutputMode {
 	case utils.OutputModeColumns:
-		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-16s %-6d %-6d %-10d %s",
+		sb.WriteString(fmt.Sprintf("%-16s %-16s %-16s %-16s %-6d %-16s %-3d %3d %s",
 			e.Node, e.Namespace, e.Pod, e.Container,
-			e.Comm, e.Pid, e.Tid, e.MountNsID, getCall(&e)))
+			e.Pid, e.Comm, e.Fd, e.Err, e.Path))
 	case utils.OutputModeCustomColumns:
 		for _, col := range params.CustomColumns {
 			switch col {
@@ -115,30 +101,14 @@ func mountsnoopTransformLine(line string) string {
 				sb.WriteString(fmt.Sprintf("%-16s", e.Container))
 			case "pid":
 				sb.WriteString(fmt.Sprintf("%-6d", e.Pid))
-			case "tid":
-				sb.WriteString(fmt.Sprintf("%-6d", e.Tid))
-			case "mnt_ns":
-				sb.WriteString(fmt.Sprintf("%-10d", e.MountNsID))
 			case "comm":
 				sb.WriteString(fmt.Sprintf("%-16s", e.Comm))
-			case "op":
-				sb.WriteString(fmt.Sprintf("%-6s", e.Operation))
-			case "ret":
-				sb.WriteString(fmt.Sprintf("%-4d", e.Retval))
-			case "lat":
-				sb.WriteString(fmt.Sprintf("%-8d", e.Latency/1000))
-			case "fs":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Fs))
-			case "src":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Source))
-			case "target":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Target))
-			case "data":
-				sb.WriteString(fmt.Sprintf("%-16s", e.Data))
-			case "flags":
-				sb.WriteString(fmt.Sprintf("%-16s", strings.Join(e.Flags, " | ")))
-			case "call":
-				sb.WriteString(fmt.Sprintf("%-16s", getCall(&e)))
+			case "fd":
+				sb.WriteString(fmt.Sprintf("%-2d", e.Fd))
+			case "err":
+				sb.WriteString(fmt.Sprintf("%-3d", e.Err))
+			case "path":
+				sb.WriteString(fmt.Sprintf("%-24s", e.Path))
 			}
 			sb.WriteRune(' ')
 		}
@@ -147,7 +117,7 @@ func mountsnoopTransformLine(line string) string {
 	return sb.String()
 }
 
-func getCustomMountsnoopColsHeader(cols []string) string {
+func getCustomOpensnoopColsHeader(cols []string) string {
 	var sb strings.Builder
 
 	for _, col := range cols {
@@ -162,30 +132,14 @@ func getCustomMountsnoopColsHeader(cols []string) string {
 			sb.WriteString(fmt.Sprintf("%-16s", "CONTAINER"))
 		case "pid":
 			sb.WriteString(fmt.Sprintf("%-6s", "PID"))
-		case "tid":
-			sb.WriteString(fmt.Sprintf("%-6s", "TID"))
-		case "mnt_ns":
-			sb.WriteString(fmt.Sprintf("%-10s", "MNT_NS"))
 		case "comm":
 			sb.WriteString(fmt.Sprintf("%-16s", "COMM"))
-		case "op":
-			sb.WriteString(fmt.Sprintf("%-6s", "OP"))
-		case "ret":
-			sb.WriteString(fmt.Sprintf("%-4s", "RET"))
-		case "lat":
-			sb.WriteString(fmt.Sprintf("%-8s", "LAT(us)"))
-		case "fs":
-			sb.WriteString(fmt.Sprintf("%-16s", "FS"))
-		case "src":
-			sb.WriteString(fmt.Sprintf("%-16s", "SRC"))
-		case "target":
-			sb.WriteString(fmt.Sprintf("%-16s", "TARGET"))
-		case "data":
-			sb.WriteString(fmt.Sprintf("%-16s", "DATA"))
-		case "flags":
-			sb.WriteString(fmt.Sprintf("%-16s", "FLAGS"))
-		case "call":
-			sb.WriteString(fmt.Sprintf("%-16s", "CALL"))
+		case "fd":
+			sb.WriteString(fmt.Sprintf("%-3s", "FD"))
+		case "err":
+			sb.WriteString(fmt.Sprintf("%-3s", "ERR"))
+		case "path":
+			sb.WriteString(fmt.Sprintf("%-24s", "PATH"))
 		}
 		sb.WriteRune(' ')
 	}
