+++ docs/guides/trace/oomkill.md
@@ -18,7 +18,7 @@ pod/test-pod created
 You can now use the gadget, but output will be empty:
 
 ```bash
-$ kubectl gadget oomkill -n oomkill-demo
+$ kubectl gadget trace oomkill -n oomkill-demo
 NODE             NAMESPACE        POD              CONTAINER        KPID   KCOMM            PAGES  TPID             TCOMM
 ```
 
@@ -59,7 +59,7 @@ You can restrain the information printed using `-o custom-columns=column0,...,co
 This command will only show the PID and command of the killed process:
 
 ```bash
-$ kubectl gadget oomkill -A -o custom-columns=kpid,kcomm
+$ kubectl gadget trace oomkill -A -o custom-columns=kpid,kcomm
 KPID   KCOMM
 15182  tail
 ```
@@ -67,7 +67,7 @@ KPID   KCOMM
 The following command is the same as default printing:
 
 ```bash
-$ kubectl gadget oomkill -A -o custom-columns=node,namespace,container,pod,kpid,kcomm,pages,tpid,tcomm
+$ kubectl gadget trace oomkill -A -o custom-columns=node,namespace,container,pod,kpid,kcomm,pages,tpid,tcomm
 NODE             NAMESPACE        CONTAINER        POD              KPID   KCOMM            PAGES  TPID   TCOMM
 minikube         oomkill-demo     test-container   test-pod         17349  tail             32768  17349  tail
 ```
@@ -77,10 +77,10 @@ minikube         oomkill-demo     test-container   test-pod         17349  tail 
 This gadget supports JSON output, for this simply use `-o json`:
 
 ```bash
-$ kubectl gadget oomkill -A -o json
+$ kubectl gadget trace oomkill -A -o json
 {"type":"normal","node":"minikube","namespace":"oomkill-demo","pod":"test-pod","container":"test-container","tpid":13416,"tcomm":"tail","kpid":13416,"kcomm":"tail","pages":32768,"mountnsid":4026532588}
 # You can use jq to make the output easier to read:
-$ kubectl gadget oomkill -A -o json | jq
+$ kubectl gadget trace oomkill -A -o json | jq
 {
   "type": "normal",
   "node": "minikube",
