+++ docs/guides/trace/open.md
@@ -1,9 +1,9 @@
 ---
-title: 'The "opensnoop" gadget'
+title: 'The "trace open" gadget'
 weight: 10
 ---
 
-The opensnoop gadget watches files that programs in pods open.
+The trace open gadget watches files that programs in pods open.
 
 Here we deploy a small demo pod "mypod":
 
@@ -11,12 +11,12 @@ Here we deploy a small demo pod "mypod":
 $ kubectl run --restart=Never -ti --image=busybox mypod -- sh -c 'while /bin/true ; do whoami ; sleep 3 ; done'
 ```
 
-Using the opensnoop gadget, we can see which processes open what files.
+Using the trace open gadget, we can see which processes open what files.
 We can simply filter for the pod "mypod" and omit specifying the node,
-thus snooping on all nodes for pod "mypod":
+thus traceing on all nodes for pod "mypod":
 
 ```bash
-$ kubectl gadget opensnoop --podname mypod
+$ kubectl gadget trace open --podname mypod
 NODE             NAMESPACE        POD              CONTAINER       PID    COMM               FD ERR PATH
 ip-10-0-30-247   default          mypod            mypod           18455  whoami              3   0 /etc/passwd
 ip-10-0-30-247   default          mypod            mypod           18521  whoami              3   0 /etc/passwd
@@ -27,7 +27,7 @@ Terminating!
 ```
 
 Seems the whoami command opens "/etc/passwd" to map the user ID to a user name.
-We can leave opensnoop by hitting Ctrl-C.
+We can leave trace open by hitting Ctrl-C.
 
 Finally, we need to clean up our pod:
