+++ docs/guides/trace/tcpconnect.md
@@ -11,7 +11,7 @@ the outgoing connections of our future pod (don't terminate it with Ctrl-C for n
 
 
 ```bash
-$ kubectl gadget tcpconnect --podname mypod
+$ kubectl gadget trace tcpconnect --podname mypod
 ```
 
 When we run the pod in a new terminal, we see the output `ok` since the public HTTP server was reached.
@@ -24,7 +24,7 @@ ok
 In our Inspektor Gadget terminal we can now see the logged connection:
 
 ```bash
-$ kubectl gadget tcpconnect --podname mypod
+$ kubectl gadget trace tcpconnect --podname mypod
 Tracing connect ... Hit Ctrl-C to end
 NODE             NAMESPACE        POD              CONTAINER       PID    COMM         IP SADDR            DADDR            DPORT
 ip-10-0-30-247   default          mypod            mypod           9386   wget         4  172.17.0.3       1.1.1.1          80
@@ -34,7 +34,7 @@ ip-10-0-30-247   default          mypod            mypod           9386   wget  
 (If the pod was started as part of a deployment, the name of the pod is not known
 in advance since random characters will be added as suffix.
 In that case, it is still possible to trace the connections. We would just
-use `kubectl gadget tcpconnect --selector key=value` to filter the pods by
+use `kubectl gadget trace tcpconnect --selector key=value` to filter the pods by
 labels instead of names.)
 
 There was a HTTP redirect to HTTPS, so we need to allow both ports for our pod.
@@ -88,7 +88,7 @@ Switching to the Inspektor Gadget terminal, we see the same connections again
 (but now with a new PID since it's a new pod):
 
 ```bash
-$ kubectl gadget tcpconnect --podname mypod  # (still running in old terminal)
+$ kubectl gadget trace tcpconnect --podname mypod  # (still running in old terminal)
 NODE             NAMESPACE        POD              CONTAINER       PID    COMM         IP SADDR            DADDR            DPORT
 ip-10-0-30-247   default          mypod            mypod           9386                wget         4  10.2.232.47      1.1.1.1          80  # (previous output)
 ip-10-0-30-247   default          mypod            mypod           9386                wget         4  10.2.232.47      1.1.1.1          443 # (previous output)
@@ -111,7 +111,7 @@ connection the pod wanted to make in the last line. Since connecting to port 80 
 there is no redirect visible to port 443:
 
 ```bash
-$ kubectl gadget tcpconnect --podname mypod  # (still running in old terminal)
+$ kubectl gadget trace tcpconnect --podname mypod  # (still running in old terminal)
 NODE             NAMESPACE        POD              CONTAINER       PID    COMM         IP SADDR            DADDR            DPORT
 ip-10-0-30-247   default          mypod            mypod           9386   wget         4  10.2.232.47      1.1.1.1          80  # (previous output)
 ip-10-0-30-247   default          mypod            mypod           9386   wget         4  10.2.232.47      1.1.1.1          443 # (previous output)
